Flytt nye klasser til passende mapper
