+++ src/main/kotlin/no/nav/fo/veilarbdialog/dialog/exceptions/FantIkkeDialogTrådException.kt
@@ -1,3 +1,3 @@
-package no.nav.fo.veilarbdialog.dialog
+package no.nav.fo.veilarbdialog.dialog.exceptions
 
 class FantIkkeDialogTrådException(val dialogId: String): Exception("Fant ikke dialog med id: $dialogId")
