+++ src/main/kotlin/no/nav/fo/veilarbdialog/dialog/exceptions/UgyldigDialogInputException.kt
@@ -1,3 +1,3 @@
-package no.nav.fo.veilarbdialog.dialog
+package no.nav.fo.veilarbdialog.dialog.exceptions
 
 class UgyldigDialogInputException(message: String): Exception(message)
