+++ src/main/kotlin/no/nav/fo/veilarbdialog/dialog/opprett/NyDialogFraBruker.kt
@@ -1,16 +1,18 @@
-package no.nav.fo.veilarbdialog.dialog
+package no.nav.fo.veilarbdialog.dialog.opprett
 
 import no.nav.common.types.identer.AktorId
 import no.nav.common.types.identer.Fnr
 import no.nav.fo.veilarbdialog.domain.AvsenderType
 
-class NyMeldingFraVeileder(
+data class NyDialogFraBruker(
     override val tekst: String,
-    override val dialogId: Long,
+    override val overskrift: String,
+    override val aktivitetId: String?,
     override val fnr: Fnr,
     override val aktorId: AktorId,
     override val avsenderId: String,
-): NyMelding {
+    val venterPaaSvarFraNav: Boolean,
+): NyDialog {
     override val avsenderType: AvsenderType
-        get() = AvsenderType.VEILEDER
+        get() = AvsenderType.BRUKER
 }
