+++ src/main/kotlin/no/nav/fo/veilarbdialog/dialog/opprett/NyMeldingFraBruker.kt
@@ -1,17 +1,17 @@
-package no.nav.fo.veilarbdialog.dialog
+package no.nav.fo.veilarbdialog.dialog.opprett
 
 import no.nav.common.types.identer.AktorId
 import no.nav.common.types.identer.Fnr
 import no.nav.fo.veilarbdialog.domain.AvsenderType
 
-class NyEskaleringsVarselDialog(
-    override val overskrift: String,
-    override val aktivitetId: String?,
+class NyMeldingFraBruker(
     override val tekst: String,
+    override val dialogId: Long,
     override val fnr: Fnr,
     override val aktorId: AktorId,
     override val avsenderId: String,
-): NyDialog {
+): NyMelding {
     override val avsenderType: AvsenderType
-        get() = AvsenderType.VEILEDER
+        get() = AvsenderType.BRUKER
 }
+
