+++ src/main/kotlin/no/nav/fo/veilarbdialog/dialog/opprett/NyMeldingFraVeileder.kt
@@ -1,19 +1,16 @@
-package no.nav.fo.veilarbdialog.dialog
+package no.nav.fo.veilarbdialog.dialog.opprett
 
 import no.nav.common.types.identer.AktorId
 import no.nav.common.types.identer.Fnr
 import no.nav.fo.veilarbdialog.domain.AvsenderType
 
-data class NyDialogFraVeileder(
+class NyMeldingFraVeileder(
     override val tekst: String,
-    override val overskrift: String,
-    override val aktivitetId: String?,
-    val venterPaaSvarFraNav: Boolean,
-    val venterPaaSvarFraBruker: Boolean,
+    override val dialogId: Long,
     override val fnr: Fnr,
     override val aktorId: AktorId,
     override val avsenderId: String,
-): NyDialog {
+): NyMelding {
     override val avsenderType: AvsenderType
         get() = AvsenderType.VEILEDER
 }
